pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "db-to-pd"
version = "0.0.1"
authors = [
  { name="Ishan Sital", email="irsital@gmail.com" },
]
description = "Enables SSH database connections and table interaction via Pandas DataFrames."
readme = "README.md"
requires-python = ">=3.7"
license = {file = "LICENSE"}
keywords = ["pandas", "postgres", "database", "ssh", "tunnel", "port-forwarding"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "psycopg2-binary~=2.9.7",
    "sshtunnel~=0.4.0",
    "pandas~=2.0.3",
    "python-dotenv~=1.0.0",
    "openpyxl~=3.1.2"
]

[project.urls]
"Homepage" = "https://github.com/irsital/db-to-pd"
"Bug Tracker" = "https://github.com/irsital/db-to-pd/issues"
